+++ lib/protocol.ex
@@ -13,8 +13,19 @@ defmodule Moar.Protocol do
   """
   @spec implements!(any(), module()) :: any()
   def implements!(x, protocol) do
-    if protocol.impl_for(x) == nil,
-      do: raise("Expected #{inspect(x)} to implement protocol #{inspect(protocol)}"),
-      else: x
+    if implements?(x, protocol),
+      do: x,
+      else: raise("Expected #{inspect(x)} to implement protocol #{inspect(protocol)}")
   end
+
+  @doc """
+  Returns true if `x` implements `protocol`.
+
+  ```elixir
+  iex> Moar.Protocol.implements?(~D[2000-01-02], String.Chars)
+  true
+  ```
+  """
+  @spec implements?(any(), module()) :: boolean()
+  def implements?(x, protocol), do: protocol.impl_for(x) != nil
 end